requirements.txt
holytools @ git+https://github.com/Somerandomguy10111/holytools
transformers==4.48.0
groq
torch
accelerate>=0.26.0
marker-pdf==1.0.0
ollama
